backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
tensorflow
pillow
python-dotenv
websockets